+++ src/buffer.h
@@ -1,6 +1,8 @@
 #ifndef __BUFFER_H__
 #define __BUFFER_H__
 
-void** adjust_buffer (void**,unsigned const,unsigned const);
+#include <stdio.h>
+
+void** adjust_buffer (void**, size_t const, size_t const);
 
 #endif /* __BUFFER_H__ */
